refactor: mv KeyServer to test-utils
